refactor(be): move updateCanCreateCourse to user module
