+++ apps/backend/apps/admin/src/group/group.service.ts
@@ -253,22 +253,6 @@ export class GroupService {
       include: includeOption
     })
   }
-
-  async updateCanCreateCourse(userId: number, canCreateCourse: boolean) {
-    return await this.prisma.user.update({
-      where: {
-        id: userId
-      },
-      data: {
-        canCreateCourse
-      },
-      select: {
-        id: true,
-        role: true,
-        canCreateCourse: true
-      }
-    })
-  }
 }
 @Injectable()
 export class InvitationService {
