+++ apps/backend/apps/admin/src/user/model/user.output.ts
@@ -1,0 +1,15 @@
+import { Field, Int } from '@nestjs/graphql'
+import { ObjectType } from '@nestjs/graphql'
+import { Role } from '@generated'
+
+@ObjectType()
+export class CanCreateCourseResult {
+  @Field(() => Int)
+  id: number
+
+  @Field(() => Role)
+  role: Role
+
+  @Field()
+  canCreateCourse: boolean
+}
