+++ apps/backend/apps/admin/src/user/user.module.ts
@@ -1,9 +1,16 @@
 import { Module } from '@nestjs/common'
-import { UserResolver } from './user.resolver'
-import { UserService } from './user.service'
+import { RolesModule } from '@libs/auth'
+import { GroupMemberResolver, UserResolver } from './user.resolver'
+import { GroupMemberService, UserService } from './user.service'
 
 @Module({
-  providers: [UserResolver, UserService],
-  exports: [UserService]
+  imports: [RolesModule],
+  providers: [
+    UserResolver,
+    UserService,
+    GroupMemberResolver,
+    GroupMemberService
+  ],
+  exports: [UserService, GroupMemberService]
 })
 export class UserModule {}
